feat(metrics): move metrics to its own endpoint
